requirements.txt
backcall==0.1.0
beautifulsoup4==4.8.0
brotlipy==0.7.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
cloudinary==1.18.1
coreapi==2.3.3
coreschema==0.0.4
cssselect==1.1.0
decorator==4.4.0
Django==2.2.6
django-cloudinary-storage==0.2.3
django-cors-headers==3.1.1
django-debug-toolbar==2.0
django-environ==0.4.5
django-extensions==2.2.1
django-filter==2.2.0
django-recaptcha==2.0.4
django-taggit==1.1.0
django-taggit-serializer==0.1.7
django-tinymce4-lite==1.7.5
djangorestframework==3.10.3
gunicorn==19.9.0
idna==2.8
ipython==7.8.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.15.1
Jinja2==2.10.1
jsmin==2.2.2
lxml==4.4.1
Markdown==3.1.1
markdown2==2.3.8
MarkupSafe==1.1.1
mock==3.0.5
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.0
prompt-toolkit==2.0.9
psycopg2==2.8.3
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.4.2
pyquery==1.4.0
pytz==2019.2
readtime==1.1.1
requests==2.22.0
scandir==1.10.0
sh==1.12.14
six==1.12.0
soupsieve==1.9.4
sqlparse==0.3.0
taggit-selectize==2.7.1
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.25.6
wcwidth==0.1.7
